feat(plotter): integrate live IMU telemetry and modularize 3D engine
- Refactored `ratatui-wireframe` by splitting the monolithic library into `model.rs` and `widget.rs`, decoupling the 3D geometry from the UI rendering logic.
- Implemented a static, labeled Global Reference Frame (X, Y, Z axes) inside the 3D canvas to visualize absolute orientation.
- Updated the `--simulate` loop to generate continuous, sine-wave driven Pitch, Yaw, and Roll telemetry.
- Wired the `WireframeWidget` directly to the live sensor state, dynamically parsing and converting IMU degree data into radians for the 3D math engine.
- Fixed the "ghost sensor" bug by explicitly flushing the OS serial input buffer after the Arduino DTR reset delay to destroy bootloader garbage bytes.
- Added the `Mock_IMU_Telemetry` Arduino sketch to `code_tests` for hardware simulation.
